# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sauropod",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.2"
  },
  "servers": [
    {
      "url": "{protocol}://{host}:{port}",
      "description": "Local server",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "The host"
        },
        "port": {
          "default": "8080",
          "description": "The port number"
        },
        "protocol": {
          "default": "http",
          "description": "The protocol"
        }
      }
    }
  ],
  "paths": {
...
    "/api/models": {
      "get": {
        "summary": "Get the list of available models",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Array_of_ModelDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/observability/logs": {
      "get": {
        "summary": "Get the logs from the system",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task": {
      "get": {
        "summary": "Get the list of tasks",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Array_of_TaskInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/int64"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/run": {
      "get": {
        "summary": "Get a list of task runs",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Array_of_TaskRunInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/run/{id}": {
      "get": {
        "summary": "Get a task run by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRun"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/{id}/run": {
      "post": {
        "summary": "Run a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Map_of_AnyValue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map_of_AnyValue"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/task/{id}/schema": {
      "get": {
        "summary": "Get the input and output JSON Schemas for a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputAndOutputSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools": {
      "get": {
        "summary": "Get the list of available tools",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Array_of_ToolDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Array_of_ModelDefinition": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ModelDefinition"
        }
      },
      "Array_of_TaskInfo": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/TaskInfo"
        }
      },
      "Array_of_TaskRunInfo": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/TaskRunInfo"
        }
      },
      "Array_of_ToolDefinition": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ToolDefinition"
        }
      },
      "Error": {
        "description": "An error message.",
        "type": "object",
        "properties": {
          "error": {
            "description": "The error message.",
            "type": "string"
          }
        },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "InputAndOutputSchema": {
        "description": "Input and output schemas.",
        "type": "object",
        "properties": {
          "inputSchema": {
            "description": "The input schema for a task.",
            "type": "object",
            "additionalProperties": true
          },
          "outputSchema": {
            "description": "The output schema for a task.",
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["inputSchema", "outputSchema"],
        "additionalProperties": false
      },
...
      "LogMessage": {
        "description": "A logged message.",
        "type": "object",
        "properties": {
          "fields": {
            "description": "Structured logging fields.",
            "type": "object",
            "additionalProperties": true
          },
          "level": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "line": {
            "description": "The line number where the log message was emitted.",
            "type": "integer",
            "format": "uint32",
            "nullable": true,
            "minimum": 0.0
          },
          "module": {
            "description": "The module where the log message was emitted.",
            "type": "string"
          },
          "timestampMs": {
            "description": "The number of milliseconds since epoch.",
            "type": "integer",
            "format": "int64"
          }
        },
        "required": ["module", "timestampMs", "fields", "level"],
        "additionalProperties": false
      },
      "LogResponse": {
        "description": "A list of logged messages.",
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/LogMessage"
        }
      },
      "Map_of_AnyValue": {
        "type": "object",
        "additionalProperties": true
      },
      "ModelDefinition": {
        "description": "A model definition.",
        "type": "object",
        "properties": {
          "name": {
            "description": "The human-friendly name of the model.",
            "type": "string"
          },
          "uri": {
            "description": "The URI of the model.",
            "type": "string"
          }
        },
        "required": ["name", "uri"],
        "additionalProperties": false
      },
...
      "Task": {
        "description": "A task is the smallest unit of work in a workflow.",
        "type": "object",
        "properties": {
          "availableToolIds": {
            "description": "The IDs of tools to make available to the LLM.",
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          },
          "inputSchema": {
            "description": "The input schema of a task.\n\n Each key in this object should be a variable name in `template`.",
            "type": "object",
            "default": {},
            "additionalProperties": true
          },
          "name": {
            "description": "The name of the task.",
            "type": "string"
          },
          "outputSchema": {
            "description": "The output schema.\n\n If unspecified the task will return an object with a single \"output\" string.",
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "template": {
            "$ref": "#/components/schemas/Template"
          }
        },
        "required": ["name", "template"],
        "additionalProperties": false
      },
...
      "TaskRun": {
        "description": "A run of a task.",
        "type": "object",
        "properties": {
          "endTimeMs": {
            "description": "The end time of the run in milliseconds since UTC epoch.",
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "id": {
            "description": "The ID of the run.",
            "type": "integer",
            "format": "int64"
          },
          "startTimeMs": {
            "description": "The start of the run in milliseconds since UTC epoch.",
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "steps": {
            "description": "The steps in the run.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Step"
            }
          }
        },
        "required": ["id", "steps", "status"],
        "additionalProperties": false
      },
...
      "int64": {
        "type": "integer",
        "format": "int64"
      },
...
    }
  }
}